# proto/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cross chanel transfer",
    "version": "1.0",
    "license": {
      "name": "Apache-2.0"
    }
  },
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/status/{idTransfer}": {
      "get": {
        "summary": "Get status transfer by id transfer",
        "operationId": "transferStatus",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/channel_transferTransferStatusResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/channel_transferErrorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "idTransfer",
            "description": "transfer ID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Transfer"
        ]
      }
    },
    "/v1/transfer/admin": {
      "post": {
        "summary": "Transfer cross chanel be admin",
        "operationId": "transferByAdmin",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/channel_transferTransferStatusResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/channel_transferErrorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/channel_transferTransferBeginAdminRequest"
            }
          }
        ],
        "tags": [
          "Transfer"
        ]
      }
    },
    "/v1/transfer/customer": {
      "post": {
        "summary": "Transfer cross chanel be customer",
        "operationId": "transferByCustomer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/channel_transferTransferStatusResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/channel_transferErrorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/channel_transferTransferBeginCustomerRequest"
            }
          }
        ],
        "tags": [
          "Transfer"
        ]
      }
    },
    "/v1/multi-transfer/admin": {
      "post": {
        "summary": "Multi transfer cross chanel be admin",
        "operationId": "multiTransferByAdmin",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/channel_transferTransferStatusResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/channel_transferErrorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/channel_transferMultiTransferBeginAdminRequest"
            }
          }
        ],
        "tags": [
          "Transfer"
        ]
      }
    },
    "/v1/multi-transfer/customer": {
      "post": {
        "summary": "Multi transfer cross chanel be customer",
        "operationId": "multiTransferByCustomer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/channel_transferTransferStatusResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/channel_transferErrorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/channel_transferMultiTransferBeginCustomerRequest"
            }
          }
        ],
        "tags": [
          "Transfer"
        ]
      }
    }
  },
  "definitions": {
    "channel_transferErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "title": "error message"
        }
      },
      "title": "ErrorResponse response with an error"
    },
    "channel_transferGeneralParams": {
      "type": "object",
      "properties": {
        "methodName": {
          "type": "string",
          "title": "method name"
        },
        "requestId": {
          "type": "string",
          "title": "request ID (may be empty)"
        },
        "chaincode": {
          "type": "string",
          "title": "chaincode name"
        },
        "channel": {
          "type": "string",
          "title": "channel name"
        },
        "nonce": {
          "type": "string",
          "title": "request nonce"
        },
        "publicKey": {
          "type": "string",
          "title": "public key of user, signed the request"
        },
        "sign": {
          "type": "string",
          "title": "request sign"
        },
        "options": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufOption"
          }
        }
      },
      "title": "GeneralParams general parameters for every request"
    },
    "channel_transferTransferBeginAdminRequest": {
      "type": "object",
      "properties": {
        "generals": {
          "$ref": "#/definitions/channel_transferGeneralParams",
          "title": "general parameters"
        },
        "idTransfer": {
          "type": "string",
          "title": "transfer ID (should be unique)"
        },
        "channelTo": {
          "type": "string",
          "title": "destination channel"
        },
        "address": {
          "type": "string",
          "title": "token's owner address"
        },
        "token": {
          "type": "string",
          "title": "tokens to transfer"
        },
        "amount": {
          "type": "string",
          "title": "amount of tokens to transfer"
        },
        "options": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufOption"
          }
        }
      },
      "title": "TransferBeginAdminRequest request for tokens transfer by platform administrator"
    },
    "channel_transferTransferBeginCustomerRequest": {
      "type": "object",
      "properties": {
        "generals": {
          "$ref": "#/definitions/channel_transferGeneralParams",
          "title": "general parameters"
        },
        "idTransfer": {
          "type": "string",
          "title": "transfer ID (should be unique)"
        },
        "channelTo": {
          "type": "string",
          "title": "destination channel"
        },
        "token": {
          "type": "string",
          "title": "Tokens to transfer"
        },
        "amount": {
          "type": "string",
          "title": "Amount of tokens to transfer"
        },
        "options": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufOption"
          }
        }
      },
      "title": "TransferBeginAdminRequest request for tokens transfer by tokens owner"
    },
    "channel_transferMultiTransferBeginAdminRequest": {
      "type": "object",
      "properties": {
        "generals": {
          "$ref": "#/definitions/channel_transferGeneralParams",
          "title": "general parameters"
        },
        "idTransfer": {
          "type": "string",
          "title": "transfer ID (should be unique)"
        },
        "channelTo": {
          "type": "string",
          "title": "destination channel"
        },
        "address": {
          "type": "string",
          "title": "token's owner address"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/channel_transferTransferItem"
          }
        },
        "options": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufOption"
          }
        }
      },
      "title": "TransferBeginAdminRequest request for tokens transfer by platform administrator"
    },
    "channel_transferMultiTransferBeginCustomerRequest": {
      "type": "object",
      "properties": {
        "generals": {
          "$ref": "#/definitions/channel_transferGeneralParams",
          "title": "general parameters"
        },
        "idTransfer": {
          "type": "string",
          "title": "transfer ID (should be unique)"
        },
        "channelTo": {
          "type": "string",
          "title": "destination channel"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/channel_transferTransferItem"
          }
        },
        "options": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufOption"
          }
        }
      },
      "title": "TransferBeginAdminRequest request for tokens transfer by tokens owner"
    },
    "channel_transferTransferStatusResponse": {
      "type": "object",
      "properties": {
        "idTransfer": {
          "type": "string",
          "title": "transfer ID"
        },
        "status": {
          "$ref": "#/definitions/channel_transferTransferStatusResponseStatus",
          "title": "transfer status"
        },
        "message": {
          "type": "string",
          "title": "error message (may be empty)"
        }
      },
      "title": "TransferStatusResponse response for TransferStatus transaction"
    },
    "channel_transferTransferStatusResponseStatus": {
      "type": "string",
      "enum": [
        "STATUS_IN_PROCESS",
        "STATUS_UNDEFINED",
        "STATUS_COMPLETED",
        "STATUS_CANCELED",
        "STATUS_ERROR"
      ],
      "default": "STATUS_IN_PROCESS"
    },
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
  }
}